cia402: merge duplicate enums; read 0x603F when a fault is detected
- Merge CiA402::OperatingMode constants namespace and enum class into a
  single enum class in CiA402Defs.hpp; CiA402Config.hpp now includes it.
  int8_t call sites use static_cast.
- Merge the three ControlWord definitions: enum class (camelCase members)
  in CiA402Defs.hpp is the single source of values; the builder helpers in
  CiA402StateMachine.hpp move to namespace ControlWords and delegate to the
  enum; EtherCAT::ControlWord becomes an alias.
- enable()/waitAtLeast() now actively read the CiA 402 error code (0x603F)
  via SDO when a fault is detected, during transitions, and on final
  failure — the log now names the fault being reset instead of just
  repeating "Resetting fault".

+++ examples/as715n_sensorless_homing.cpp
@@ -231,8 +231,8 @@ bool configureAndEnableDrive(EtherCAT::DS402Master& master, const SensorlessHomi
     }
 
     const int8_t op_mode = args.use_csv_mode
-        ? CiA402::OperatingMode::CyclicSyncVelocity
-        : CiA402::OperatingMode::CyclicSyncTorque;
+        ? static_cast<int8_t>(CiA402::OperatingMode::CyclicSyncVelocity)
+        : static_cast<int8_t>(CiA402::OperatingMode::CyclicSyncTorque);
     if (!init.drive().setOperatingMode(op_mode)) {
         TETHER_LOGE(TAG, "Failed to set operating mode {}", static_cast<int>(op_mode));
         return false;
